Draft refactoring. Moved base extractors.
