+++ fastopenapi/routers/aiohttp/async_router.py
@@ -6,7 +6,8 @@ from aiohttp import web
 from fastopenapi.core.types import Response
 from fastopenapi.openapi.ui import render_redoc_ui, render_swagger_ui
 from fastopenapi.routers.aiohttp.extractors import AioHttpRequestDataExtractor
-from fastopenapi.routers.base import BaseAdapter, RequestEnvelope
+from fastopenapi.routers.base import BaseAdapter
+from fastopenapi.routers.common import RequestEnvelope
 
 
 class AioHttpRouter(BaseAdapter):
