+++ fastopenapi/routers/falcon/sync_router.py
@@ -5,10 +5,8 @@ import falcon
 
 from fastopenapi.core.types import Response
 from fastopenapi.openapi.ui import render_redoc_ui, render_swagger_ui
-from fastopenapi.routers.base import (
-    BaseAdapter,
-    RequestEnvelope,
-)
+from fastopenapi.routers.base import BaseAdapter
+from fastopenapi.routers.common import RequestEnvelope
 from fastopenapi.routers.falcon.extractors import FalconRequestDataExtractor
