+++ fastopenapi/routers/quart/async_router.py
@@ -5,7 +5,8 @@ from quart import jsonify, request
 
 from fastopenapi.core.types import Response
 from fastopenapi.openapi.ui import render_redoc_ui, render_swagger_ui
-from fastopenapi.routers.base import BaseAdapter, RequestEnvelope
+from fastopenapi.routers.base import BaseAdapter
+from fastopenapi.routers.common import RequestEnvelope
 from fastopenapi.routers.quart.extractors import QuartRequestDataExtractor
