+++ fastopenapi/routers/sanic/async_router.py
@@ -4,7 +4,8 @@ from sanic import response
 
 from fastopenapi.core.types import Response
 from fastopenapi.openapi.ui import render_redoc_ui, render_swagger_ui
-from fastopenapi.routers.base import BaseAdapter, RequestEnvelope
+from fastopenapi.routers.base import BaseAdapter
+from fastopenapi.routers.common import RequestEnvelope
 from fastopenapi.routers.sanic.extractors import SanicRequestDataExtractor
