+++ fastopenapi/routers/starlette/async_router.py
@@ -7,7 +7,8 @@ from starlette.routing import Route
 
 from fastopenapi.core.types import Response
 from fastopenapi.openapi.ui import render_redoc_ui, render_swagger_ui
-from fastopenapi.routers.base import BaseAdapter, RequestEnvelope
+from fastopenapi.routers.base import BaseAdapter
+from fastopenapi.routers.common import RequestEnvelope
 from fastopenapi.routers.starlette.extractors import StarletteRequestDataExtractor
